+++ include/build.h
@@ -2,6 +2,6 @@
  * Author: sascha_lammers@gmx.de
  */
 #pragma once
-#define __BUILD_NUMBER "8845"
-#define __BUILD_NUMBER_INT 8845
+#define __BUILD_NUMBER "8850"
+#define __BUILD_NUMBER_INT 8850